Time decoupled automated releases from checks
This to ensure that no double decrement can happen right on the border
of reservation validity. Due to the fact that the automated release of
some reservation will only happen strictly after the validity has
expired.

Currently the extra period added is 10 seconds for normal reservations
and twice that (20 seconds) for child reservations